requirements.txt
alembic==1.11.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-tools==1.56.0
lamia-shared @ git+https://github.com/erfansahebi/lamia_shared@7b1f964bd8940b519550a97f9f1076702d84bad0
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
protobuf==4.23.3
psycopg2-binary==2.9.6
SQLAlchemy==2.0.17
typing_extensions==4.7.1
